Extract the view model into a protocol
Also made the installJetpack method accept a Blog instance
instead of the url/username/password combination.

This prepares the protocol for the upcoming Jetpack Individual Plugin case
where the installation method uses the REST API through Jetpack Proxy instead.
